requirements.txt
requests
flask
PyJWT
beautifulsoup4
waitress
flask-wtf
yk-utils
